application/Major/requirements.txt
Flask==3.0.3
torch
torchvision
xgboost
numpy
pillow
opencv-python